+++ flutter_app/lib/infrastructure/database.dart
@@ -16,7 +16,7 @@ import 'package:database_builder/database_builder.dart';
 
 class DatabaseHelper {
   static const _databaseName = "database.db";
-  static const _databaseVersion = 29;
+  static const _databaseVersion = 30;
   late Database _database;
 
   Future<void> init() async {