+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 - Fix bug with chapter chooser not disappearing.
 - Allow users to choose the old grid-style chapter chooser in settings.
-- Add an alternate verse chooser grid in settings.
+- Add verse chooser layout option (Grid / Sidebar) in settings.
 - Show full book name as a popup in the grid book chooser.
 - Show book name and chapter number at the beginning of each chapter.
+- Add words of Jesus in red option to settings.
 
 # 2.1.0 - 2026.09.15